Refactoring_of_code

+++ src/main/java/com/h2rd/refactoring/web/UserResource.java
@@ -7,6 +7,9 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.GenericEntity;
 import javax.ws.rs.core.Response;
 
+import com.h2rd.refactoring.usermanagement.User_Service;
+import com.h2rd.refactoring.usermanagement.Validating_data;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.stereotype.Repository;
@@ -20,7 +23,9 @@ public class UserResource{
 
     public UserDao userDao;
 
-    @GET
+    @Autowired
+    private User_Service user_service = new User_Service();
+    @POST
     @Path("add/")
     public Response addUser(@QueryParam("name") String name,
                             @QueryParam("email") String email,
@@ -30,12 +35,17 @@ public class UserResource{
         user.setName(name);
         user.setEmail(email);
         user.setRoles(roles);
-
-        if (userDao == null) {
-            userDao = UserDao.getUserDao();
+        ArrayList<String> validations_defined = new Validating_data().validate_User(user);
+        if (validations_defined.size() > 0) {
+            return Response.status(Response.Status.FORBIDDEN).entity(validations_defined).build();
         }
 
-        userDao.saveUser(user);
+        if (user_service.isUserExist(user)) {
+            return Response.status(Response.Status.CONFLICT).entity("User '" + user.getEmail() + "' already exist.")
+                    .build();
+        }
+
+        user_service.saveUser(user);
         return Response.ok().entity(user).build();
     }
 
@@ -50,11 +60,16 @@ public class UserResource{
         user.setEmail(email);
         user.setRoles(roles);
 
-        if (userDao == null) {
-            userDao = UserDao.getUserDao();
+        ArrayList<String> validations_defined = new Validating_data().validate_User(user);
+        if (validations_defined.size() > 0) {
+            return Response.status(Response.Status.FORBIDDEN).entity(validations_defined).build();
+        }
+        if (!user_service.isUserExist(user)) {
+            return Response.status(Response.Status.CONFLICT).entity("User '" + user.getEmail() + "' already exist.")
+                    .build();
         }
 
-        userDao.updateUser(user);
+        user_service.updateUser(user);
         return Response.ok().entity(user).build();
     }
 
@@ -68,8 +83,13 @@ public class UserResource{
         user.setEmail(email);
         user.setRoles(roles);
 
-        if (userDao == null) {
-            userDao = UserDao.getUserDao();
+        ArrayList<String> validations_defined = new Validating_data().validate_User(user);
+        if (validations_defined.size() > 0) {
+            return Response.status(Response.Status.FORBIDDEN).entity(validations_defined).build();
+        }
+        if (user_service.isUserExist(user)) {
+            return Response.status(Response.Status.CONFLICT).entity("User '" + user.getEmail() + "' already exist.")
+                    .build();
         }
 
         userDao.deleteUser(user);
@@ -79,14 +99,10 @@ public class UserResource{
     @GET
     @Path("find/")
     public Response getUsers() {
-    	
-        ApplicationContext context = new ClassPathXmlApplicationContext(new String[] {
-    		"classpath:/application-config.xml"	
-    	});
-    	userDao = context.getBean(UserDao.class);
-    	List<User> users = userDao.getUsers();
+
+    	List<User> users = user_service.getUsers();
     	if (users == null) {
-    		users = new ArrayList<User>();
+            return Response.status(Response.Status.FORBIDDEN).entity("User not found.").build();
     	}
 
         GenericEntity<List<User>> usersEntity = new GenericEntity<List<User>>(users) {};
@@ -95,13 +111,11 @@ public class UserResource{
 
     @GET
     @Path("search/")
-    public Response findUser(@QueryParam("name") String name) {
-
-        if (userDao == null) {
-            userDao = UserDao.getUserDao();
+    public Response findUser(@QueryParam("email") String email) {
+        User user = user_service.findUserByEmail(email);
+        if (user == null) {
+            return Response.status(Response.Status.FORBIDDEN).entity("User not found.").build();
         }
-
-        User user = userDao.findUser(name);
         return Response.ok().entity(user).build();
     }
 }
